+++ packages/web3-utils/CHANGELOG.md
@@ -42,3 +42,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added and exported three reusable utility functions: `pollTillDefined`, `rejectIfTimeout` and `rejectIfConditionAtInterval` which are useful when dealing with promises that involves polling, rejecting after timeout or rejecting if a condition was met when calling repeatably at every time intervals.
 
 ## [Unreleased]
+
+### Added
+
+-   Export a new function `uuidV4` that generates a random v4 Uuid (#5373).
+
+### Fixed
+
+-   Use Uuid for the response id, to fix the issue "Responses get mixed up due to conflicting payload IDs" (#5373).